Spec updates, fixes for running random & multi
