+++ spec/ruby/core/math/atanh_spec.rb
@@ -1,5 +1,4 @@
 require_relative '../../spec_helper'
-require_relative '../../fixtures/math/common'
 require_relative '../../shared/math/atanh'
 
 describe "Math.atanh" do
@@ -8,7 +7,8 @@ describe "Math.atanh" do
 end
 
 describe "Math#atanh" do
-  it_behaves_like :math_atanh_private, :atanh
-  it_behaves_like :math_atanh_base, :atanh, IncludesMath.new
-  it_behaves_like :math_atanh_no_complex, :atanh, IncludesMath.new
+  math = Class.new { include ::Math }.new
+  it_behaves_like :math_atanh_private   , :atanh
+  it_behaves_like :math_atanh_base      , :atanh, math
+  it_behaves_like :math_atanh_no_complex, :atanh, math
 end
