--- spec/ruby/fixtures/math/common.rb
@@ -1,3 +1,0 @@
-class IncludesMath
-  include Math
-end
