+++ spec/ruby/library/complex/math/atanh_spec.rb
@@ -1,14 +1,14 @@
 require 'complex'
 require_relative '../../../spec_helper'
-require_relative '../../../fixtures/math/common'
 require_relative '../../../shared/math/atanh'
 require_relative 'shared/atanh'
 
 describe "Math#atanh" do
-  it_behaves_like :math_atanh_base, :atanh, IncludesMath.new
-  it_behaves_like :complex_math_atanh_complex, :atanh, IncludesMath.new
+  math = Class.new { include ::CMath }.new
+  it_behaves_like :math_atanh_base, :atanh, math
+  it_behaves_like :complex_math_atanh_complex, :atanh, math
 
-  it_behaves_like :math_atanh_private, :atanh, IncludesMath.new
+  it_behaves_like :math_atanh_private, :atanh, math
 end
 
 describe "Math.atanh" do
